+++ lib/240326/repository/comment_repository.dart
@@ -1,4 +1,4 @@
-import '../data_source/comments.dart';
+import '../data_source/comment.dart';
 
 abstract interface class CommentRepository {
   Future<List<Comment>> getComments(int postId);
