+++ lib/240326/repository/comment_repository_impl.dart
@@ -1,4 +1,4 @@
-import 'package:learn_dart_together/240326/data_source/comments.dart';
+import 'package:learn_dart_together/240326/data_source/comment.dart';
 import 'package:learn_dart_together/240326/model/comment_api.dart';
 import 'package:learn_dart_together/240326/repository/comment_repository.dart';
